+++ src/dft/analysis.cpp
@@ -2,13 +2,14 @@
 
 #include "storm-dft/parser/DFTJsonParser.h"
 #include "storm-dft/builder/ExplicitDFTModelBuilder.h"
-#include "storm-dft/storage/dft/DFTIsomorphism.h"
+#include "storm-dft/storage/DFTIsomorphism.h"
 
+template<typename ValueType> using ExplicitDFTModelBuilder = storm::dft::builder::ExplicitDFTModelBuilder<ValueType>;
 
 // Thin wrapper for DFT analysis
 template<typename ValueType>
-std::vector<ValueType> analyzeDFT(storm::storage::DFT<ValueType> const& dft, std::vector<std::shared_ptr<storm::logic::Formula const>> const& properties, bool symred, bool allowModularisation, storm::utility::RelevantEvents const& relevantEvents, bool allowDCForRelevant) {
-    typename storm::modelchecker::DFTModelChecker<ValueType>::dft_results dftResults = storm::api::analyzeDFT(dft, properties, symred, allowModularisation, relevantEvents, allowDCForRelevant, 0.0, storm::builder::ApproximationHeuristic::DEPTH, false);
+std::vector<ValueType> analyzeDFT(storm::dft::storage::DFT<ValueType> const& dft, std::vector<std::shared_ptr<storm::logic::Formula const>> const& properties, bool symred, bool allowModularisation, storm::dft::utility::RelevantEvents const& relevantEvents, bool allowDCForRelevant) {
+    typename storm::dft::modelchecker::DFTModelChecker<ValueType>::dft_results dftResults = storm::dft::api::analyzeDFT(dft, properties, symred, allowModularisation, relevantEvents, allowDCForRelevant, 0.0, storm::dft::builder::ApproximationHeuristic::DEPTH, false);
 
     std::vector<ValueType> results;
     for (auto result : dftResults) {
@@ -17,26 +18,54 @@ std::vector<ValueType> analyzeDFT(storm::storage::DFT<ValueType> const& dft, std
     return results;
 }
 
+// Thin wrapper for building state space from DFT
+template<typename ValueType>
+std::shared_ptr<storm::models::sparse::Model<ValueType>> buildModel(storm::dft::storage::DFT<ValueType> const& dft, storm::dft::storage::DFTIndependentSymmetries const& symmetries, storm::dft::utility::RelevantEvents const& relevantEvents, bool allowDCForRelevant) {
+    dft.setRelevantEvents(relevantEvents, allowDCForRelevant);
+    storm::dft::builder::ExplicitDFTModelBuilder<ValueType> builder(dft, symmetries);
+    builder.buildModel(0, 0.0);
+    return builder.getModel();
+}
 
 // Define python bindings
 void define_analysis(py::module& m) {
-    // RelevantEvents
-    py::class_<storm::utility::RelevantEvents, std::shared_ptr<storm::utility::RelevantEvents>>(m, "RelevantEvents", "Relevant events which should be observed")
-        .def("is_relevant", &storm::utility::RelevantEvents::isRelevant, "Check whether the given name is a relevant event", py::arg("name"))
+
+    py::enum_<storm::dft::builder::ApproximationHeuristic>(m, "ApproximationHeuristic", "Heuristic for selecting states to explore next")
+        .value("DEPTH", storm::dft::builder::ApproximationHeuristic::DEPTH)
+        .value("PROBABILITY", storm::dft::builder::ApproximationHeuristic::PROBABILITY)
+        .value("BOUNDDIFFERENCE", storm::dft::builder::ApproximationHeuristic::BOUNDDIFFERENCE)
     ;
 
-    m.def("analyze_dft", &analyzeDFT<double>, "Analyze the DFT", py::arg("dft"), py::arg("properties"), py::arg("symred")=true, py::arg("allow_modularisation")=false, py::arg("relevant_events")=storm::utility::RelevantEvents(), py::arg("allow_dc_for_relevant")=false);
-    m.def("analyze_parametric_dft", &analyzeDFT<storm::RationalFunction>, "Analyze the parametric DFT", py::arg("dft"), py::arg("properties"), py::arg("symred")=true, py::arg("allow_modularisation")=false, py::arg("relevant_events")=storm::utility::RelevantEvents(), py::arg("allow_dc_for_relevant")=false);
-
-    m.def("transform_dft", &storm::api::applyTransformations<double>, "Apply transformations on DFT", py::arg("dft"), py::arg("unique_constant_be"), py::arg("binary_fdeps"));
-    m.def("transform_parametric_dft", &storm::api::applyTransformations<storm::RationalFunction>, "Apply transformations on parametric DFT", py::arg("dft"), py::arg("unique_constant_be"), py::arg("binary_fdeps"));
-
-    m.def("compute_dependency_conflicts", &storm::api::computeDependencyConflicts<double>, "Set conflicts between FDEPs. Is used in analysis.", py::arg("dft"), py::arg("use_smt") = false, py::arg("solver_timeout") = 0);
-    m.def("compute_dependency_conflicts_parametric", &storm::api::computeDependencyConflicts<storm::RationalFunction>, "Set conflicts between FDEPs. Is used in analysis.", py::arg("dft"), py::arg("use_smt") = false, py::arg("solver_timeout") = 0);
-
-    m.def("is_well_formed", &storm::api::isWellFormed<double>, "Check whether DFT is well-formed.", py::arg("dft"), py::arg("check_valid_for_analysis") = true);
-    m.def("is_well_formed_parametric", &storm::api::isWellFormed<storm::RationalFunction>, "Check whether parametric DFT is well-formed.", py::arg("dft"), py::arg("check_valid_for_analysis") = true);
-
-    m.def("compute_relevant_events", &storm::api::computeRelevantEvents<double>, "Compute relevant event ids from properties and additional relevant names", py::arg("dft"), py::arg("properties"), py::arg("additional_relevant_names") = std::vector<std::string>());
-    m.def("compute_relevant_events_parametric", &storm::api::computeRelevantEvents<storm::RationalFunction>, "Compute relevant event ids from properties and additional relevant names", py::arg("dft"), py::arg("properties"), py::arg("additional_relevant_names") = std::vector<std::string>());
+    // RelevantEvents
+    py::class_<storm::dft::utility::RelevantEvents, std::shared_ptr<storm::dft::utility::RelevantEvents>>(m, "RelevantEvents", "Relevant events which should be observed")
+        .def(py::init<>(), "Create empty list of relevant events")
+        .def("is_relevant", &storm::dft::utility::RelevantEvents::isRelevant, "Check whether the given name is a relevant event", py::arg("name"))
+    ;
 }
+
+template<typename ValueType>
+void define_analysis_typed(py::module& m, std::string const& vt_suffix) {
+
+    py::class_<ExplicitDFTModelBuilder<ValueType>, std::shared_ptr<ExplicitDFTModelBuilder<ValueType>>>(m, ("ExplicitDFTModelBuilder"+vt_suffix).c_str(), "Builder to generate explicit model from DFT")
+        .def(py::init<storm::dft::storage::DFT<ValueType> const&, storm::dft::storage::DFTIndependentSymmetries const&>(), "Constructor", py::arg("dft"), py::arg("symmetries"))
+        .def("build", &ExplicitDFTModelBuilder<ValueType>::buildModel, "Build state space of model", py::arg("iteration"), py::arg("approximation_threshold")=0.0, py::arg("approximation_heuristic")=storm::dft::builder::ApproximationHeuristic::DEPTH)
+        .def("get_model", &ExplicitDFTModelBuilder<ValueType>::getModel, "Get complete model")
+        .def("get_partial_model", &ExplicitDFTModelBuilder<ValueType>::getModelApproximation, "Get partial model", py::arg("lower_bound"), py::arg("expected_time"))
+    ;
+
+    m.def(("_analyze_dft"+vt_suffix).c_str(), &analyzeDFT<ValueType>, "Analyze the DFT", py::arg("dft"), py::arg("properties"), py::arg("symred")=true, py::arg("allow_modularisation")=false, py::arg("relevant_events")=storm::dft::utility::RelevantEvents(), py::arg("allow_dc_for_relevant")=false);
+
+    m.def(("_build_model"+vt_suffix).c_str(), &buildModel<ValueType>, "Build state-space model (CTMC or MA) for DFT", py::arg("dft"), py::arg("symmetries"), py::arg("relevant_events")=storm::dft::utility::RelevantEvents(), py::arg("allow_dc_for_relevant")=false);
+
+    m.def(("_transform_dft"+vt_suffix).c_str(), &storm::dft::api::applyTransformations<ValueType>, "Apply transformations on DFT", py::arg("dft"), py::arg("unique_constant_be"), py::arg("binary_fdeps"), py::arg("exponential_distributions"));
+
+    m.def(("_compute_dependency_conflicts"+vt_suffix).c_str(), &storm::dft::api::computeDependencyConflicts<ValueType>, "Set conflicts between FDEPs. Is used in analysis.", py::arg("dft"), py::arg("use_smt") = false, py::arg("solver_timeout") = 0);
+
+    m.def(("_is_well_formed"+vt_suffix).c_str(), &storm::dft::api::isWellFormed<ValueType>, "Check whether DFT is well-formed.", py::arg("dft"), py::arg("check_valid_for_analysis") = true);
+
+    m.def(("_compute_relevant_events"+vt_suffix).c_str(), &storm::dft::api::computeRelevantEvents<ValueType>, "Compute relevant event ids from properties and additional relevant names", py::arg("dft"), py::arg("properties"), py::arg("additional_relevant_names") = std::vector<std::string>());
+}
+
+
+template void define_analysis_typed<double>(py::module& m, std::string const& vt_suffix);
+template void define_analysis_typed<storm::RationalFunction>(py::module& m, std::string const& vt_suffix);